# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Developer Portal (ADP) API
  version: 1.0.0-alpha
  description: |
    REST API for AI agent governance, audit trails, and documentation.
    Supports both PostgreSQL and SQLite backends.

    ## Overview

    ADP provides two core capabilities for AI coding agents:

    1. **Decision Audit Trail** -- every agent decision is logged with reasoning, confidence, and policy outcome.
    2. **Documentation Engine** -- auto-generates session summaries, risk reports, and pattern analysis from decision records.

    ## Response Formats

    The REST API (`adp-server`, port 8080) uses envelope wrappers for all responses:

    - **Single object**: `{"data": <object>}`
    - **Created object** (201): `{"data": <object>}`
    - **List**: `{"items": [...], "total": 10, "limit": 50, "offset": 0}`
    - **Error**: `{"error": "message", "code": "bad_request", "details": null}`

    ## Sidecar vs REST API

    The commit endpoints (`/v1/commits/*`) exist in two places with **different response formats**:

    - **REST API** (this spec, port 8080): Wraps responses in `{"data": {...}}` envelopes.
    - **MCP HTTP sidecar** (port 8081): Returns top-level JSON without envelopes. Git hooks expect the sidecar format.

    See the commit endpoint descriptions for details on both formats.

    ## Authentication

    ADP supports three authentication methods:

...

    The first registered user is automatically granted the admin role.
...

    ## Database Backends

...

...
servers:
...
paths:
...
  /v1/admin/users:
    get:
      operationId: listUsers
      summary: List all users (admin only)
      description: Returns a paginated list of all user accounts. Requires admin role.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/users/{id}:
    get:
      operationId: getUser
      summary: Get a user by ID (admin only)
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UserProfile"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateUser
      summary: Update a user's role, status, or name (admin only)
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid role or status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: disableUser
      summary: Disable a user account (admin only)
      description: Soft-deletes the user by setting status to disabled. Cannot disable your own account.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User disabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User disabled"
        "400":
          description: Cannot disable own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication via locally-generated tokens (POST /v1/auth/login)
        or external JWKS provider. Required in production (PostgreSQL mode).
        Optional in SQLite development mode.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Shared API key authentication. Set via ADP_API_KEY environment variable.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: User ID
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip for pagination

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
            code: "not_found"
            details: null
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "insufficient permissions"
            code: "forbidden"
            details: null
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - bad_request
            - not_found
            - unauthorized
            - forbidden
            - internal_error
            - conflict
        details:
          description: Additional error context (nullable)
          oneOf:
            - type: object
            - type: "null"

    # -------------------------------------------------------------------------
    # Sessions
    # -------------------------------------------------------------------------
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
          description: User role (first user is auto-admin)
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      description: All fields are optional. Only provided fields are updated.
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        status:
          type: string
          enum: [active, disabled]
    UserListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer